introduce exceptions module in Typer
